Adds a tokenizer. (#10)
Because our language is LL(1), the tokenizer is implemented as a generator. The parser will then be able to tokenize while parsing is occurring.

A slight implementation detail: Tokens have different kinds, which are strings instead of enums. This introduces a bit of development risk, but ultimately yields simpler reading code.